Support HDR subtraction mode
This mode is mostly used for background removal use case.
One frame is captured with illumination, one without, and both are
subtracted by the sensor itself, producing a frame with only the
illuminated subject.
This requires multi context support as it needs 2 internal frames for 1
output frame, and new blanking checks to make sure the sensor has enough
time to process the whole pipe.

Signed-off-by: Benjamin Mugnier <benjamin.mugnier@foss.st.com>