Make hblank control read-only
Libcamera requires hblank control to be settable at stream, or purely
read-only.
The vd55g1 handles a range of hblank but they cannot be changed during
stream, confusing libcamera.
As much as I despise doing userspace specific code in a driver, making the
stack aware of this hardware specificity require a patch in libcamera
and in the userspace api. So let's limit this for now, and come back
later once patches are done.

Signed-off-by: Benjamin Mugnier <benjamin.mugnier@foss.st.com>